requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.7.0
CacheControl==0.14.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
Django==3.2
django-cors-headers==4.4.0
djangorestframework==3.15.1
firebase-admin==6.5.0
Flask==3.0.2
Flask-Cors==4.0.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==22.0.0
httplib2==0.22.0
idna==3.6
importlib_metadata==8.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jwt==1.3.1
kombu==5.3.7
lxml==5.2.2
MarkupSafe==2.1.5
msgpack==1.0.8
npm==0.1.1
optional-django==0.1.0
packaging==24.1
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.2
zipp==3.19.2
